Fix infinite recursion bug in wait_for_finish.
In #52, a bug was introduced to OutputSchema.any_failed which caused
calls to OutputSchema.wait_for_finish() to recurse infinitely.
wait_for_finish called any_failed for the is_failed check, which in turn
called wait_for_finish.

For small files, the transformation step finished before the maximum
recursion limit was reached, but large files resulted in a crash.